PIANO-4879: make some changes to ELC to get it to work with piano.

+++ Classes/ELCImagePicker/ELCAsset.h
@@ -25,7 +25,7 @@
 @property (nonatomic, strong) ALAsset *asset;
 @property (nonatomic, weak) id<ELCAssetDelegate> parent;
 @property (nonatomic, assign) BOOL selected;
-@property (nonatomic,assign) int index;
+@property (nonatomic,assign) NSInteger index;
 
 - (id)initWithAsset:(ALAsset *)asset;
 - (NSComparisonResult)compareWithIndex:(ELCAsset *)_ass;
